moved tests and ui datatypes
