+++ modules/docs/arrow-docs/build.gradle
@@ -23,6 +23,7 @@ dependencies {
     compile project(':arrow-mtl-data')
     compile project(':arrow-free-data')
     compile project(':arrow-test')
+    compile project(':arrow-ui')
     compile project(':arrow-free-extensions')
     compile project(':arrow-generic')
     compile project(':arrow-query-language')
