+++ modules/docs/arrow-docs/docs/docs/arrow/data/store/README.md
@@ -19,7 +19,7 @@ If we think in a component oriented fashion when building user interfaces, this 
 This structure is also a [`Comonad`]({{ '/docs/arrow/typeclasses/comonad' | relative_url }}) because it represents a lazy unfolding of all possible states of our user interface.
 
 ```kotlin:ank
-import arrow.data.*
+import arrow.ui.*
 
 val store = Store(0) { "The current value is: $it" }
 store.extract()
@@ -56,7 +56,7 @@ upperCaseStore.extract()
 
 ```kotlin:ank:replace
 import arrow.reflect.*
-import arrow.data.*
+import arrow.ui.*
 import arrow.core.*
 
 DataType(Store::class).tcMarkdownList()
