+++ modules/free/arrow-free-data/src/test/kotlin/arrow/free/TrampolineTest.kt
@@ -1,8 +1,5 @@
-package arrow.data
+package arrow.free
 
-import arrow.free.Trampoline
-import arrow.free.TrampolineF
-import arrow.free.runT
 import io.kotlintest.runner.junit4.KotlinTestRunner
 import org.junit.runner.RunWith
 import arrow.test.UnitSpec
