+++ modules/mtl/arrow-mtl-data/src/test/kotlin/arrow/mtl/OptionTTest.kt
@@ -7,7 +7,10 @@ import arrow.core.Option
 import arrow.core.Some
 import arrow.core.ForNonEmptyList
 import arrow.core.NonEmptyList
+import arrow.core.ForId
+import arrow.core.Id
 import arrow.core.extensions.const.divisible.divisible
+import arrow.core.extensions.id.monad.monad
 import arrow.core.extensions.monoid
 import arrow.effects.ForIO
 import arrow.effects.IO
@@ -17,6 +20,9 @@ import arrow.effects.extensions.optiont.async.async
 import arrow.effects.typeclasses.seconds
 import arrow.core.extensions.nonemptylist.monad.monad
 import arrow.core.extensions.option.monad.monad
+import arrow.core.fix
+import arrow.core.value
+import arrow.mtl.extensions.ComposedFunctorFilter
 import arrow.mtl.extensions.optiont.applicative.applicative
 import arrow.mtl.extensions.optiont.divisible.divisible
 import arrow.mtl.extensions.optiont.monoidK.monoidK
@@ -33,12 +39,17 @@ import arrow.test.laws.SemigroupKLaws
 import arrow.test.laws.TraverseFilterLaws
 import arrow.typeclasses.Const
 import arrow.typeclasses.Eq
+import arrow.typeclasses.NestedType
 import arrow.typeclasses.Monad
 import arrow.typeclasses.const
+import arrow.typeclasses.nest
+import arrow.typeclasses.unnest
 import arrow.typeclasses.value
 import io.kotlintest.properties.forAll
 import io.kotlintest.runner.junit4.KotlinTestRunner
 import org.junit.runner.RunWith
+
+typealias OptionTNel = Kind<OptionTPartialOf<ForNonEmptyList>, Int>
 
 @RunWith(KotlinTestRunner::class)
 class OptionTTest : UnitSpec() {
@@ -63,6 +74,17 @@ class OptionTTest : UnitSpec() {
 
   init {
 
+    val EQ_OPTIONT_ID_NEL: Eq<NestedType<OptionTPartialOf<ForId>, OptionTPartialOf<ForNonEmptyList>, Int>> =
+      Eq { a, b ->
+        a.unnest().value().value().fold(
+          { b.unnest().value().value().isEmpty() },
+          { optionA: OptionTNel ->
+            b.unnest().value().value().fix().fold(
+              { false },
+              { it.value() == optionA.value() })
+          })
+      }
+
     testLaws(
       AsyncLaws.laws(OptionT.async(IO.async()), IOEQ(), IOEitherEQ()),
 
@@ -70,6 +92,12 @@ class OptionTTest : UnitSpec() {
         OptionT.semigroupK(Option.monad()),
         OptionT.applicative(Option.monad()),
         EQ()),
+
+      FunctorFilterLaws.laws(
+        ComposedFunctorFilter(OptionT.functorFilter(Id.monad()),
+          OptionT.functorFilter(NonEmptyList.monad())),
+        { OptionT.just(Id.monad(), OptionT.just(NonEmptyList.monad(), it)).nest() },
+        EQ_OPTIONT_ID_NEL),
 
       MonoidKLaws.laws(
         OptionT.monoidK(Option.monad()),
