+++ modules/ui/arrow-ui-data/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     testRuntime("org.junit.vintage:junit-vintage-engine:$jUnitVintageVersion")
     testCompile "io.kotlintest:kotlintest-runner-junit4:$kotlinTestVersion"
     testCompile project(':arrow-test')
+    testCompile project(':arrow-ui')
 }
 
 apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
