+++ modules/ui/arrow-ui-data/src/main/kotlin/arrow/ui/Sum.kt
@@ -1,4 +1,4 @@
-package arrow.data
+package arrow.ui
 
 import arrow.Kind
 import arrow.higherkind
