+++ modules/ui/arrow-ui-data/src/test/kotlin/arrow/ui/MooreTest.kt
@@ -1,7 +1,7 @@
-package arrow.data
+package arrow.ui
 
 import arrow.core.Id
-import arrow.data.extensions.moore.comonad.comonad
+import arrow.ui.extensions.moore.comonad.comonad
 import arrow.test.UnitSpec
 import arrow.test.laws.ComonadLaws
 import arrow.typeclasses.Eq
