+++ modules/ui/arrow-ui-data/src/test/kotlin/arrow/ui/SumTest.kt
@@ -1,4 +1,4 @@
-package arrow.data
+package arrow.ui
 
 import arrow.Kind
 import arrow.core.ForId
@@ -12,10 +12,10 @@ import arrow.core.extensions.id.functor.functor
 import arrow.core.extensions.id.hash.hash
 import arrow.core.extensions.monoid
 import arrow.core.fix
-import arrow.data.extensions.sum.comonad.comonad
-import arrow.data.extensions.sum.divisible.divisible
-import arrow.data.extensions.sum.eq.eq
-import arrow.data.extensions.sum.hash.hash
+import arrow.ui.extensions.sum.comonad.comonad
+import arrow.ui.extensions.sum.divisible.divisible
+import arrow.ui.extensions.sum.eq.eq
+import arrow.ui.extensions.sum.hash.hash
 import arrow.test.UnitSpec
 import arrow.test.laws.ComonadLaws
 import arrow.test.laws.DivisibleLaws
@@ -23,10 +23,9 @@ import arrow.test.laws.HashLaws
 import arrow.typeclasses.Const
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Hash
-import arrow.typeclasses.fix
 import arrow.typeclasses.value
-import io.kotlintest.shouldBe
 import io.kotlintest.runner.junit4.KotlinTestRunner
+import io.kotlintest.shouldBe
 import org.junit.runner.RunWith
 
 @RunWith(KotlinTestRunner::class)
