+++ modules/ui/arrow-ui/src/main/kotlin/arrow/ui/extensions/store.kt
@@ -1,9 +1,9 @@
-package arrow.data.extensions
+package arrow.ui.extensions
 
 import arrow.Kind
-import arrow.data.Store
-import arrow.data.StorePartialOf
-import arrow.data.fix
+import arrow.ui.Store
+import arrow.ui.StorePartialOf
+import arrow.ui.fix
 import arrow.extension
 import arrow.typeclasses.Comonad
 import arrow.typeclasses.Functor
